Logging | BasicAuth | Documentation

+++ facade_app/util/pagingStoreController.py
@@ -2,15 +2,17 @@ import json, os, datetime
 from pymongo import MongoClient
 import pymongo
 
+LOG_LEVEL = os.environ["LOG_LEVEL"]
+
 def connectToMongoPaging(page_store_time=1200):
     
-    # Provide the mongodb atlas url to connect python to mongodb using pymongo
+    # Provide the mongodb connector url to connect 
     CONNECTION_STRING = os.environ["MONGO_DB_URL"]
 
-    # Create a connection using MongoClient. You can import MongoClient or use pymongo.MongoClient
+    # Create a connection to the database
     client = MongoClient(CONNECTION_STRING)
 
-    # Create the database for our example (we will use the same database throughout the tutorial
+    # Create the database or connect to the collection if it already exists
     db = client['Paging']
     if(not 'Pages' in db.list_collection_names()):
         col = db['Pages']
@@ -28,7 +30,7 @@ def storePage(page, page_id, page_store_time=1200):
             with open(f"./pages/{page_id}",'w') as outFile:
                 json.dump(page, outFile, indent="")
         except IOError:
-            print(IOError)
+            print("Error writing page to disk")
             return f"Error writing page with ID {page_id}", 403
 
 def getPage(page_id, remove=False):
@@ -50,7 +52,8 @@ def clearPages(page_store_time, dir_to_search="./pages/"):
     else:
         for dirpath, dirnames, filenames in os.walk(dir_to_search):
             for file in filenames:
-                curpath = os.path.join(dirpath, file)
-                file_modified = datetime.datetime.fromtimestamp(os.path.getmtime(curpath))
-                if datetime.datetime.now() - file_modified > datetime.timedelta(seconds=page_store_time):
-                    os.remove(curpath)
+                if(file != "init.cfg"):
+                    curpath = os.path.join(dirpath, file)
+                    file_modified = datetime.datetime.fromtimestamp(os.path.getmtime(curpath))
+                    if datetime.datetime.now() - file_modified > datetime.timedelta(seconds=page_store_time):
+                        os.remove(curpath)
